requirements.txt
django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
psycopg2-binary==2.9.3
celery==5.2.6
django-celery-results==2.3.1
gunicorn==20.1.0
boto3==1.22.11